// src/telegram-webapp.d.ts
interface WebAppUser {
    /*
     * Уникальный идентификатор пользователя или бота. Это число может содержать более 32 значащих бит, и некоторые языки программирования могут испытывать трудности/скрытые дефекты при его интерпретации. Оно содержит не более 52 значащих бит, поэтому для хранения этого идентификатора можно использовать 64-битное целое число или тип с плавающей точкой двойной точности.
     */
    id: number;
    /*
     * *Необязательно*. True, если этот пользователь — бот. Возвращается только в поле ***получатель***.
     */
    is_bot?: boolean;
    /*
     * Имя пользователя или бота.
     */
    first_name: string;
    /*
     * *Необязательно*. Фамилия пользователя или бота.
     */
    last_name?: string;
    /*
     * *Необязательно*. Имя пользователя или бота.
     */
    username?: string;
    /*
     * *Необязательно*. Тег ***[IETF language](https://en.wikipedia.org/wiki/IETF_language_tag)*** языка пользователя. Возвращается только в поле пользователя.
     */
    language_code?: string;
    /*
     * *Необязательно*. True, если этот пользователь — пользователь Telegram Premium.
     */
    is_premium?: true;
    /*
     * *Необязательно*. True, если пользователь добавил бота в меню вложений.
     */
    added_to_attachment_menu?: true;
    /*
     * *Необязательно*. True, если пользователь разрешил боту отправлять ему сообщения.
     */
    allows_write_to_pm?: true;
    /*
     * *Необязательно*. URL-адрес фотографии профиля пользователя. Фотография может быть в формате .jpeg или .svg.
     */
    photo_url?: string;
}

/*
 * Этот объект представляет собой чат.
 */
interface WebAppChat {
    /*
     * Уникальный идентификатор этого чата. Это число может иметь более 32 значимых бит, и некоторые языки программирования могут испытывать трудности/скрытые дефекты при его интерпретации. Однако оно имеет не более 52 значимых бит, поэтому для хранения этого идентификатора можно использовать 64-битное целое число со знаком или число с плавающей точкой двойной точности.
     */
    id: number;
    /*
     * Тип чата: «группа (group)», «супергруппа (supergroup)» или «канал (channel)».
     */
    type: 'group' | 'supergroup' | 'channel';
    /*
     * Название чата
     */
    title: string;
    /*
     * *Необязательно*. Имя пользователя чата
     */
    username?: string;
    /*
     * *Необязательно*. URL-адрес фотографии чата. Фотография может быть в формате .jpeg или .svg. Возвращается только для мини-приложений, запущенных из меню вложений.
     */
    photo_url?: string;
}

/*
 * Этот объект содержит данные, передаваемые в мини-приложение при его открытии. Он пуст, если мини-приложение было запущено с помощью ***[кнопки клавиатуры (keyboard button)](https://core.telegram.org/bots/webapps#keyboard-button-mini-apps)*** или из ***[встроенного режима (inline mode)](https://core.telegram.org/bots/webapps#inline-mode-mini-apps)***.
 */
export interface WebAppInitData {
    /*
     * *Необязательно*. Уникальный идентификатор сеанса мини-приложения, необходимый для отправки сообщений через метод **answerWebAppQuery**.
     */
    query_id?: string;
    /*
     * *Необязательно*. Объект, содержащий данные о текущем пользователе.
     */
    user?: WebAppUser;
    /*
     * *Необязательно*. Объект, содержащий данные о собеседнике текущего пользователя в чате, где бот был запущен через меню вложений. Возвращается только для приватных чатов и только для мини-приложений, запущенных через меню вложений.
     */
    receiver?: WebAppUser;
    /*
     * *Необязательно*. Объект, содержащий данные о чате, где бот был запущен через меню вложений. Возвращается для супергрупп, каналов и групповых чатов – только для мини-приложений, запущенных через меню вложений.
     */
    chat?: WebAppChat;
    /*
     * *Необязательно*. Тип чата, из которого было открыто мини-приложение. Может быть «отправитель» для личного чата с пользователем, открывающим ссылку, «личный (private)», «группа (group)», «супергруппа (supergroup)» или «канал (channel)». Возвращается только для мини-приложений, запущенных по прямым ссылкам.
     */
    chat_type?: 'sender' | 'private' | 'group' | 'supergroup' | 'channel';
    /*
     * *Необязательно*. Глобальный идентификатор, уникальный для чата, из которого было открыто мини-приложение. Возвращается только для мини-приложений, запущенных по прямой ссылке.
     */
    chat_instance?: string;
    /*
     * *Optional*. The value of the `startattach` parameter, passed ***[via link](https://core.telegram.org/bots/webapps#adding-bots-to-the-attachment-menu)***. Only returned for Mini Apps when launched from the attachment menu via link.
     *
     * Значение параметра `start_param` также будет передано в GET-параметр `tgWebAppStartParam`, поэтому мини-приложение сможет сразу загрузить правильный интерфейс.
     */
    start_param?: string;
    /*
     * *Необязательно*. Время в секундах, по истечении которого сообщение может быть отправлено через метод **answerWebAppQuery**.
     */
    can_send_after?: number;
    /*
     * Время Unix на момент открытия формы.
     */
    auth_date: number;
    /*
     * Хеш всех переданных параметров, который бот-сервер может использовать для ***[проверки их валидности](https://core.telegram.org/bots/webapps#validating-data-received-via-the-mini-app)***.
     */
    hash: string;
    /*
     * Подпись всех переданных параметров (кроме *хеша*), которую третья сторона может использовать для ***[проверки их действительности](https://core.telegram.org/bots/webapps#validating-data-for-third-party-use)***.
     */
    signature?: string;
}